+++ VirtualH89/Src/HardSectoredDisk.cpp
@@ -11,18 +11,18 @@
 
 #include <cstring>
 
-HardSectoredDisk::HardSectoredDisk(const char *name)
+HardSectoredDisk::HardSectoredDisk(const char* name)
 {
-    FILE *file;
+    FILE* file;
 
     initialized_m = true;
     memset(rawImage_m, 0, maxHeads_c * bytesPerTrack_c * maxTracksPerSide_c);
 
     if ((file = fopen(name, "r")) != NULL)
     {
-        int readCount;
-        int trk = 0;
-        int head = 0;
+        unsigned long readCount;
+        int           trk  = 0;
+        int           head = 0;
 
         do
         {
@@ -39,7 +39,7 @@ HardSectoredDisk::HardSectoredDisk(const char *name)
             }
 
             if ((readCount =
-                        fread(&rawImage_m[head][trk][0], bytesPerTrack_c, 1, file)) == 1)
+                     fread(&rawImage_m[head][trk][0], bytesPerTrack_c, 1, file)) == 1)
             {
                 trk++;
             }
@@ -52,7 +52,6 @@ HardSectoredDisk::HardSectoredDisk(const char *name)
                 sides_m, tracks_m);
         fclose(file);
     }
-
     else
     {
         debugss(ssFloppyDisk, ERROR, "%s: unable to open file - %s\n", __FUNCTION__, name);
@@ -72,8 +71,8 @@ HardSectoredDisk::HardSectoredDisk()
     // clear disk
     bzero(rawImage_m, maxHeads_c * bytesPerTrack_c * maxTracksPerSide_c);
 
-    tracks_m = maxTracksPerSide_c;
-    sides_m  = maxHeads_c;
+    tracks_m      = maxTracksPerSide_c;
+    sides_m       = maxHeads_c;
 
     initialized_m = true;
     setWriteProtect(false);
@@ -84,10 +83,11 @@ HardSectoredDisk::~HardSectoredDisk()
 
 }
 
-bool HardSectoredDisk::readData(BYTE          side,
-                                BYTE          track,
-                                unsigned int  pos,
-                                BYTE&         data)
+bool
+HardSectoredDisk::readData(BYTE          side,
+                           BYTE          track,
+                           unsigned long pos,
+                           BYTE&         data)
 {
     debugss(ssFloppyDisk, ALL, "%s: maxTrack (%d) tracks_m(%d)\n", __FUNCTION__, maxTrack_m,
             tracks_m);
@@ -107,7 +107,6 @@ bool HardSectoredDisk::readData(BYTE          side,
                     __FUNCTION__, track, track / 2);
             track /= 2;
         }
-
         else if ((maxTrack_m == 40) && (tracks_m == 80))
         {
             debugss(ssFloppyDisk, INFO, "%s: max - 40 trk_m - 80\n", __FUNCTION__);
@@ -119,20 +118,18 @@ bool HardSectoredDisk::readData(BYTE          side,
     {
         if ((track < maxTracksPerSide_c) && (pos < bytesPerTrack_c))
         {
-            debugss(ssFloppyDisk, INFO, "%s: side(%d) track(%d) pos(%d) = %d\n", __FUNCTION__,
+            debugss(ssFloppyDisk, INFO, "%s: side(%d) track(%d) pos(%lu) = %d\n", __FUNCTION__,
                     side, track, pos, rawImage_m[side][track][pos]);
             data = rawImage_m[side][track][pos];
             return true;
         }
-
         else
         {
-            debugss(ssFloppyDisk, ERROR, "%s: range error: track(%d) pos(%d)\n",
+            debugss(ssFloppyDisk, ERROR, "%s: range error: track(%d) pos(%lu)\n",
                     __FUNCTION__, track, pos);
             return false;
         }
     }
-
     else
     {
         debugss(ssFloppyDisk, ERROR, "%s: disk not initialized\n", __FUNCTION__);
@@ -140,10 +137,11 @@ bool HardSectoredDisk::readData(BYTE          side,
     }
 }
 
-bool HardSectoredDisk::writeData(BYTE         side,
-                                 BYTE         track,
-                                 unsigned int pos,
-                                 BYTE         data)
+bool
+HardSectoredDisk::writeData(BYTE          side,
+                            BYTE          track,
+                            unsigned long pos,
+                            BYTE          data)
 {
     debugss(ssFloppyDisk, ALL, "%s: maxTrack (%d) tracks_m(%d)\n", __FUNCTION__,
             maxTrack_m, tracks_m);
@@ -162,7 +160,6 @@ bool HardSectoredDisk::writeData(BYTE         side,
             debugss(ssFloppyDisk, INFO, "%s: max - 80 trk_m - 40\n", __FUNCTION__);
             track /= 2;
         }
-
         else if ((maxTrack_m == 40) && (tracks_m == 80))
         {
             debugss(ssFloppyDisk, INFO, "%s: max - 40 trk_m - 80\n", __FUNCTION__);
@@ -175,19 +172,19 @@ bool HardSectoredDisk::writeData(BYTE         side,
         if ((track < maxTracksPerSide_c) && (pos < bytesPerTrack_c))
         {
             rawImage_m[side][track][pos] = data;
-            debugss(ssFloppyDisk, ALL, "%s: side (%d) track(%d) pos(%d) = %d\n", __FUNCTION__, side, track,
+            debugss(ssFloppyDisk, ALL, "%s: side (%d) track(%d) pos(%lu) = %d\n", __FUNCTION__,
+                    side,
+                    track,
                     pos, rawImage_m[side][track][pos]);
             return true;
         }
-
         else
         {
-            debugss(ssFloppyDisk, ERROR, "%s: Out of Range - track(%d) pos(%d)\n",
+            debugss(ssFloppyDisk, ERROR, "%s: Out of Range - track(%d) pos(%lu)\n",
                     __FUNCTION__, track, pos);
             return false;
         }
     }
-
     else
     {
         debugss(ssFloppyDisk, ERROR, "%s: disk not initialized\n", __FUNCTION__);
@@ -196,16 +193,16 @@ bool HardSectoredDisk::writeData(BYTE         side,
 
 }
 
-void HardSectoredDisk::getControlInfo(unsigned int  pos,
-                                      bool&         hole,
-                                      bool&         writeProtect)
+void
+HardSectoredDisk::getControlInfo(unsigned long pos,
+                                 bool&         hole,
+                                 bool&         writeProtect)
 {
     if (initialized_m)
     {
         hole         = defaultHoleStatus(pos);
         writeProtect = checkWriteProtect();
     }
-
     else
     {
         // No disk in drive so hole light is on.
@@ -216,9 +213,10 @@ void HardSectoredDisk::getControlInfo(unsigned int  pos,
     }
 }
 
-bool HardSectoredDisk::defaultHoleStatus(unsigned int pos)
+bool
+HardSectoredDisk::defaultHoleStatus(unsigned long pos)
 {
-    debugss(ssFloppyDisk, ALL, "%s: pos = %d ", __FUNCTION__, pos);
+    debugss(ssFloppyDisk, ALL, "%s: pos = %lu ", __FUNCTION__, pos);
 
     // check for a sector hole
     if ((pos % 320) < 64)
@@ -239,26 +237,28 @@ bool HardSectoredDisk::defaultHoleStatus(unsigned int pos)
     return false;
 }
 
-void HardSectoredDisk::dump()
+void
+HardSectoredDisk::dump()
 {
     debugss(ssFloppyDisk, ALL, "%s: Not Implemented\n", __FUNCTION__);
 }
 
-void HardSectoredDisk::eject(const char *name)
+void
+HardSectoredDisk::eject(const char* name)
 {
-    FILE *file;
+    FILE* file;
     debugss(ssFloppyDisk, ALL, "%s: Save: %s\n", __FUNCTION__, name);
 
     if ((file = fopen(name, "wb")) != NULL)
     {
-        int readCount;
+        unsigned long readCount;
 
         for (int head = 0; head < sides_m; head++)
         {
             for (int track = 0; track < tracks_m; track++)
             {
                 if ((readCount =
-                            fwrite(&rawImage_m[head][track][0], bytesPerTrack_c, 1, file)) != 1)
+                         fwrite(&rawImage_m[head][track][0], bytesPerTrack_c, 1, file)) != 1)
                 {
                     debugss(ssFloppyDisk, ERROR, "%s: Unable to save file: %s head: %d track: %d\n",
                             __FUNCTION__, name, head, track);
@@ -268,18 +268,18 @@ void HardSectoredDisk::eject(const char *name)
 
         fclose(file);
     }
-
     else
     {
         debugss(ssFloppyDisk, WARNING, "%s: unable to save file - %s\n", __FUNCTION__, name);
     }
 }
 
-bool HardSectoredDisk::readSectorData(BYTE   side,
-                                      BYTE   track,
-                                      BYTE   sector,
-                                      WORD   pos,
-                                      BYTE&  data)
+bool
+HardSectoredDisk::readSectorData(BYTE  side,
+                                 BYTE  track,
+                                 BYTE  sector,
+                                 WORD  pos,
+                                 BYTE& data)
 {
     return false;
 }
