+++ VirtualH89/Src/HardSectoredDisk.h
@@ -9,7 +9,6 @@
 #ifndef HARDSECTOREDDISK_H_
 #define HARDSECTOREDDISK_H_
 
-#include "h89Types.h"
 #include "FloppyDisk.h"
 
 class DiskData;
@@ -24,35 +23,35 @@ class DiskData;
 class HardSectoredDisk: public FloppyDisk
 {
   public:
-    HardSectoredDisk(const char *name);
+    HardSectoredDisk(const char* name);
     HardSectoredDisk();
     virtual ~HardSectoredDisk();
 
-    virtual bool readData(BYTE side, BYTE track, unsigned int pos, BYTE& data);
-    virtual bool writeData(BYTE side, BYTE track, unsigned int pos, BYTE data);
-    virtual void getControlInfo(unsigned int pos, bool& hole, bool& writeProtect);
+    virtual bool readData(BYTE side, BYTE track, unsigned long pos, BYTE& data);
+    virtual bool writeData(BYTE side, BYTE track, unsigned long pos, BYTE data);
+    virtual void getControlInfo(unsigned long pos, bool& hole, bool& writeProtect);
     void dump();
 
     virtual bool readSectorData(BYTE side, BYTE track, BYTE sector, WORD pos,
                                 BYTE& data);
-    virtual void eject(const char *name);
+    virtual void eject(const char* name);
 
   private:
     // Hard sectored disks on Heath, only support single density (FM encoding)
-    static const unsigned int bytesPerTrack_c    = 3200;
+    static const unsigned int bytesPerTrack_c = 3200;
 
     /// \todo make the tracks (and sides) runtime set, so that double sided and 80 track
     /// hard-sectored disks can be used
     static const unsigned int maxHeads_c         = 2;
     static const unsigned int maxTracksPerSide_c = 80;
 
-    BYTE rawImage_m[maxHeads_c][maxTracksPerSide_c][bytesPerTrack_c];
+    BYTE                      rawImage_m[maxHeads_c][maxTracksPerSide_c][bytesPerTrack_c];
 
-    bool initialized_m    = false;
-    unsigned int tracks_m = 80;
-    unsigned int sides_m  = 2;
+    bool                      initialized_m = false;
+    unsigned int              tracks_m      = 80;
+    unsigned int              sides_m       = 2;
 
-    bool defaultHoleStatus(unsigned int pos);
+    bool defaultHoleStatus(unsigned long pos);
 };
 
 #endif // HARDSECTOREDDISK_H_
