+++ VirtualH89/Src/h-17-1.cpp
@@ -4,10 +4,8 @@
 /// \author Mark Garlanger
 ///
 
-#include <strings.h>
-
 #include "h-17-1.h"
-#include "WallClock.h"
+
 #include "logger.h"
 #include "FloppyDisk.h"
 
@@ -22,7 +20,8 @@ H_17_1::~H_17_1()
 }
 
 
-void H_17_1::insertDisk(FloppyDisk *disk)
+void
+H_17_1::insertDisk(FloppyDisk* disk)
 {
     disk_m = disk;
 
@@ -33,11 +32,12 @@ void H_17_1::insertDisk(FloppyDisk *disk)
     }
 }
 
-void H_17_1::getControlInfo(unsigned int pos, bool& hole, bool& trackZero, bool& writeProtect)
+void
+H_17_1::getControlInfo(unsigned long pos, bool& hole, bool& trackZero, bool& writeProtect)
 {
     // Track info comes from the drive, the hole and write-protect is determined
     // by the actual disk
-    debugss(ssH17_1, INFO, "%s - pos: %d\n", __FUNCTION__, pos);
+    debugss(ssH17_1, INFO, "%s - pos: %ld\n", __FUNCTION__, pos);
 
     trackZero = (track_m == 0);
 
@@ -45,16 +45,16 @@ void H_17_1::getControlInfo(unsigned int pos, bool& hole, bool& trackZero, bool&
     {
         disk_m->getControlInfo(pos, hole, writeProtect);
     }
-
     else
     {
         debugss(ssH17_1, INFO, "%s no disk_m\n", __FUNCTION__);
-        hole = true;
+        hole         = true;
         writeProtect = false;
     }
 }
 
-void H_17_1::step(bool direction)
+void
+H_17_1::step(bool direction)
 {
     if (direction)
     {
@@ -65,7 +65,6 @@ void H_17_1::step(bool direction)
 
         debugss(ssH17_1, INFO, "%s - in(up) (%d)\n", __FUNCTION__, track_m);
     }
-
     else
     {
         if (track_m)
@@ -77,50 +76,51 @@ void H_17_1::step(bool direction)
     }
 }
 
-void H_17_1::selectSide(BYTE side)
+void
+H_17_1::selectSide(BYTE side)
 {
     // Since the H-17-1 is single-sided, this is a NOP.
 }
 
-BYTE H_17_1::readData(unsigned int pos)
+BYTE
+H_17_1::readData(unsigned long pos)
 {
     BYTE data = 0;
 
     if ((disk_m) && (disk_m->readData(head_c, track_m, pos, data)))
     {
-        debugss(ssH17_1, INFO, "%s: read passed - pos(%d) data(%d)\n", __FUNCTION__, pos, data);
+        debugss(ssH17_1, INFO, "%s: read passed - pos(%lu) data(%d)\n", __FUNCTION__, pos, data);
     }
-
     else
     {
-        debugss(ssH17_1, WARNING, "%s: read failed - pos(%d)\n", __FUNCTION__, pos);
+        debugss(ssH17_1, WARNING, "%s: read failed - pos(%lu)\n", __FUNCTION__, pos);
     }
 
     return data;
 }
 
-void H_17_1::writeData(unsigned int pos, BYTE data)
+void
+H_17_1::writeData(unsigned long pos, BYTE data)
 {
     if ((disk_m) && (!disk_m->writeData(head_c, track_m, pos, data)))
     {
-        debugss(ssH17_1, WARNING, "%s: pos(%d)\n", __FUNCTION__, pos);
+        debugss(ssH17_1, WARNING, "%s: pos(%lu)\n", __FUNCTION__, pos);
     }
 }
 
-BYTE H_17_1::readSectorData(BYTE sector, unsigned int pos)
+BYTE
+H_17_1::readSectorData(BYTE sector, unsigned long pos)
 {
     BYTE data = 0;
 
     if ((disk_m) && (disk_m->readSectorData(0, track_m, sector, pos, data)))
     {
-        debugss(ssH17_1, INFO, "%s: read passed - pos(%d) data(%d)\n", __FUNCTION__, pos, data);
+        debugss(ssH17_1, INFO, "%s: read passed - pos(%lu) data(%d)\n", __FUNCTION__, pos, data);
     }
-
     else
     {
-        debugss(ssH17_1, WARNING, "%s: read failed - pos(%d)\n", __FUNCTION__, pos);
+        debugss(ssH17_1, WARNING, "%s: read failed - pos(%lu)\n", __FUNCTION__, pos);
     }
 
     return data;
 }
-
