+++ VirtualH89/Src/h-17-1.h
@@ -7,9 +7,7 @@
 #ifndef H_17_1_H_
 #define H_17_1_H_
 
-#include "config.h"
 
-#include "h89Types.h"
 #include "DiskDrive.h"
 
 ///
@@ -27,16 +25,16 @@ class H_17_1: public DiskDrive
     H_17_1();
     virtual ~H_17_1();
 
-    void getControlInfo(unsigned int pos, bool& hole, bool& trackZero, bool& writeProtect);
+    void getControlInfo(unsigned long pos, bool& hole, bool& trackZero, bool& writeProtect);
 
     void step(bool direction);
     void selectSide(BYTE side);
 
-    BYTE readData(unsigned int pos);
-    void writeData(unsigned int pos, BYTE data);
+    BYTE readData(unsigned long pos);
+    void writeData(unsigned long pos, BYTE data);
 
-    virtual BYTE readSectorData(BYTE sector, unsigned int pos);
-    void insertDisk(FloppyDisk *disk);
+    virtual BYTE readSectorData(BYTE sector, unsigned long pos);
+    void insertDisk(FloppyDisk* disk);
 
 
   private:
