+++ VirtualH89/Src/h-17-4.cpp
@@ -5,8 +5,6 @@
 /// \author Mark Garlanger
 ///
 
-#include <strings.h>
-
 #include "h-17-4.h"
 
 #include "logger.h"
@@ -14,7 +12,7 @@
 
 
 H_17_4::H_17_4(): side_m(0),
-    track_m(0)
+                  track_m(0)
 {
     numTracks_m = maxTracks_c;
 
@@ -25,7 +23,8 @@ H_17_4::~H_17_4()
 
 }
 
-void H_17_4::insertDisk(FloppyDisk *disk)
+void
+H_17_4::insertDisk(FloppyDisk* disk)
 {
     disk_m = disk;
 
@@ -36,12 +35,13 @@ void H_17_4::insertDisk(FloppyDisk *disk)
     }
 }
 
-void H_17_4::getControlInfo(unsigned int pos,
-                            bool& hole,
-                            bool& trackZero,
-                            bool& writeProtect)
+void
+H_17_4::getControlInfo(unsigned long pos,
+                       bool&         hole,
+                       bool&         trackZero,
+                       bool&         writeProtect)
 {
-    debugss(ssH17_4, INFO, "%s - pos: %d\n", __FUNCTION__, pos);
+    debugss(ssH17_4, INFO, "%s - pos: %ld\n", __FUNCTION__, pos);
 
     trackZero = (track_m == 0);
 
@@ -49,16 +49,16 @@ void H_17_4::getControlInfo(unsigned int pos,
     {
         disk_m->getControlInfo(pos, hole, writeProtect);
     }
-
     else
     {
         debugss(ssH17_4, INFO, "%s no disk_m\n", __FUNCTION__);
-        hole = true;
+        hole         = true;
         writeProtect = false;
     }
 }
 
-void H_17_4::step(bool direction)
+void
+H_17_4::step(bool direction)
 {
     if (direction)
     {
@@ -69,7 +69,6 @@ void H_17_4::step(bool direction)
 
         debugss(ssH17_4, WARNING, "%s - in(up) (%d)\n", __FUNCTION__, track_m);
     }
-
     else
     {
         if (track_m)
@@ -81,7 +80,8 @@ void H_17_4::step(bool direction)
     }
 }
 
-void H_17_4::selectSide(BYTE side)
+void
+H_17_4::selectSide(BYTE side)
 {
     if (side == side_m)
     {
@@ -94,7 +94,6 @@ void H_17_4::selectSide(BYTE side)
     {
         side_m = side;
     }
-
     else
     {
         debugss(ssH17, ERROR, "%s: Invalid Side: %d\n", __FUNCTION__, side);
@@ -102,45 +101,46 @@ void H_17_4::selectSide(BYTE side)
 }
 
 
-BYTE H_17_4::readData(unsigned int pos)
+BYTE
+H_17_4::readData(unsigned long pos)
 {
     BYTE data = 0;
 
     if ((disk_m) && (disk_m->readData(side_m, track_m, pos, data)))
     {
-        debugss(ssH17_4, INFO, "%s: read passed - pos(%d) data(%d)\n", __FUNCTION__, pos, data);
+        debugss(ssH17_4, INFO, "%s: read passed - pos(%lu) data(%d)\n", __FUNCTION__, pos, data);
     }
-
     else
     {
-        debugss(ssH17_4, WARNING, "%s: read failed - pos(%d)\n", __FUNCTION__, pos);
+        debugss(ssH17_4, WARNING, "%s: read failed - pos(%lu)\n", __FUNCTION__, pos);
     }
 
     return data;
 }
 
-void H_17_4::writeData(unsigned int pos, BYTE data)
+void
+H_17_4::writeData(unsigned long pos, BYTE data)
 {
-    debugss(ssH17_4, INFO, "%s: pos(%d) data(%d)\n", __FUNCTION__, pos, data);
+    debugss(ssH17_4, INFO, "%s: pos(%lu) data(%d)\n", __FUNCTION__, pos, data);
 
     if ((disk_m) && (!disk_m->writeData(side_m, track_m, pos, data)))
     {
-        debugss(ssH17_4, WARNING, "%s: pos(%d)\n", __FUNCTION__, pos);
+        debugss(ssH17_4, WARNING, "%s: pos(%lu)\n", __FUNCTION__, pos);
     }
 }
 
-BYTE H_17_4::readSectorData(BYTE sector, unsigned int pos)
+BYTE
+H_17_4::readSectorData(BYTE sector, unsigned long pos)
 {
     BYTE data = 0;
 
     if ((disk_m) && (disk_m->readSectorData(side_m, track_m, sector, pos, data)))
     {
-        debugss(ssH17_4, INFO, "%s: read passed - pos(%d) data(%d)\n", __FUNCTION__, pos, data);
+        debugss(ssH17_4, INFO, "%s: read passed - pos(%lu) data(%d)\n", __FUNCTION__, pos, data);
     }
-
     else
     {
-        debugss(ssH17_4, WARNING, "%s: read failed - pos(%d)\n", __FUNCTION__, pos);
+        debugss(ssH17_4, WARNING, "%s: read failed - pos(%lu)\n", __FUNCTION__, pos);
     }
 
     return data;
